5/1/2023 Changes

+++ Final Project/Main.cpp
@@ -16,8 +16,11 @@ int main(void) {
 	card deck[deckSize];
 
 	char inputChar = 'a';
+	char dummy;
 	char fileName[20];
 	char scan[12];
+	
+	char playerNames[4][20] = { NULL, NULL, NULL, NULL };
 
 	int i;
 	int j;
@@ -27,6 +30,8 @@ int main(void) {
 	int discards;
 	int draws;
 
+	int* scores;
+
 	bool valid = false;
 
 	card drawnCard;
@@ -35,11 +40,11 @@ int main(void) {
 	
 	card* temp = NULL;
 	card* center = NULL;
-	card* players[4] = {NULL, NULL, NULL, NULL};
+	card* players[4] = { NULL, NULL, NULL, NULL };
 
 
-	//input number of players
-	printf("Let's play a game of DOS\nInput answers as (color number) with # = 11\n");
+	// inputs the number of players
+	printf("Let's play a game of DOS\n\nInput your cards using the format 'color number'\nEnter 11 as the number for Wild # Cards\n\n");
 	while (inputChar != 'q') {
 		printf("Enter the number of players: ");
 		scanf("%d", &playerNum);
@@ -54,7 +59,7 @@ int main(void) {
 		}
 	}
 	
-	//choose starting deck
+	// chooses starting deck
 	while ((inputChar != '1') && (inputChar != '2')) {
 		printf("\nPress 1 to shuffle the DOS deck or 2 to load a deck from a file: ");
 		scanf(" %c", &inputChar);
@@ -86,7 +91,24 @@ int main(void) {
 		}
 	}
 	
-	//deal out seven alternating cards to each player
+	// allocate memory to integer scores
+	scores = (int*)malloc(playerNum * sizeof(int));
+	for (i = 0; i < playerNum; i++) {
+		scores[i] = 0;
+	}
+
+	//get the player names
+	for (i = 0; i < playerNum; i++) {
+		printf("\nPlayer %d Enter your name: ", i + 1);
+		if (i == 0) {
+			scanf("%c", &dummy);
+		}
+		fgets(playerNames[i], 20, stdin);
+		playerNames[i][strlen(playerNames[i])] = '\0';
+		
+	}
+
+	// deals out seven alternating cards to each player
 	for (i = 0; i < 7; i++) {
 		for (j = 0; j < playerNum; j++) {
 			drawnCard = dealCard(deck, deckSize);
@@ -97,7 +119,7 @@ int main(void) {
 	}
 	drawnCard = dealCard(deck, deckSize);
 
-	//initalize variables for playCard, centerCard, and drawnCard
+	// initalizes variables for playCard, centerCard, and drawnCard
 	playCard[0].pt = NULL;
 	playCard[1].pt = NULL;
 	centerCard.pt = NULL;
@@ -107,14 +129,15 @@ int main(void) {
 	strcpy(centerCard.action, "none");
 	strcpy(drawnCard.action, "none");
 
+	/* italizes the number of center cards to zero before drawing 
+	and choose a random starting player */
 	centerNum = 0;
 	playerIndex = rand() % playerNum;
 
-	// game loop
+	//TODO add multiple rounds
 	//TODO reshuffle and deal new cards
-	//TODO create more functions
-	// 
-	//add up to two cards to the center
+
+	// adds up to two cards to the center
 	for (i = centerNum; i < 2; i++) {
 		drawnCard = dealCard(deck, deckSize);
 		center = addCard(center, drawnCard);
@@ -123,55 +146,68 @@ int main(void) {
 
 	while (true) {
 		
-		
 		//TODO add player names
-		//print the first center row and player hand
-		printGame(players, center, playerIndex);
+		// prints the first center row and player hand
+		printGame(players, center, playerNames, playerIndex);
 
-		//prompt the current player to do a valid action
-		//TODO add multiple plays 
+		// initalzes the number of bonus cards to play and deal to zero
 		discards = 0;
 		draws = 0;
-
-		printf("Enter 'draw' or 'play' to take an action: ");
-		scanf("%s", scan);
+		
+		// prompt the player to draw or play a card 
 		while (true) {
+			printf("Enter 'draw' or 'play' to take an action: ");
+			scanf("%s", scan);
+
+			// draw a card
 			if (strcmp(scan, "draw") == 0) {
-				//printf("You drew!!!");
 				drawnCard = dealCard(deck, deckSize);
 				players[playerIndex] = addCard(players[playerIndex], drawnCard);
-				printGame(players, center, playerIndex);
+				strcpy(scan, "continue");
 				break;
-
 			}
+
+			// obtain user input to make the first play
 			else if (strcmp(scan, "play") == 0) {
 				printf("Choose a card on the center row: ");
 				scanf("%s %d", &centerCard.color, &centerCard.value);
 				printf("Enter a card to play: ");
 				scanf("%s %d", &playCard[0].color, &playCard[0].value);
 				playCard[1].value = 0;
-				//checks if the play is valid
-				if (checkPlay(center, players[playerIndex], centerCard, playCard)) {
-					//removes cards from player hand and center row,  
+
+				// if the center card is a Wild # card, 
+				if (centerCard.value == 11) {
+					printf("Would you like to add another card? (Yes/No):");
+					scanf("%s", scan);
+				}
+
+				// check if the single match is valid 
+				if (checkPlay(center, players[playerIndex], centerCard, playCard) && (strcmp(scan, "Yes") != 0 && strcmp(scan, "yes") != 0)) {
+					
+					// play the card from the player's hand and check for bonuses
 					players[playerIndex] = deleteCard(players[playerIndex], playCard[0]);
 					center = deleteCard(center, centerCard);
 					centerNum--;
-					printGame(players, center, playerIndex);
+					printGame(players, center, playerNames, playerIndex);
 					if (strcmp(centerCard.color, playCard[0].color) == 0) {
 						discards++;
 					}
 					break;
 				}
-				//if not, prompts for an additional card
+
+				/* if the single match is invalid or the user wishes to play two cards on a wildcard,
+				prompt for an additional card */
 				else if (playCard[0].value < centerCard.value) {
 					printf("Enter another card to play: ");
 					scanf("%s %d", &playCard[1].color, &playCard[1].value);
+					
+					// if the double match is valid, play the cards and check for bonuses 
 					if (checkPlay(center, players[playerIndex], centerCard, playCard)) {
 						players[playerIndex] = deleteCard(players[playerIndex], playCard[0]);
 						players[playerIndex] = deleteCard(players[playerIndex], playCard[1]);
 						center = deleteCard(center, centerCard);
 						centerNum--;
-						printGame(players, center, playerIndex);
+						printGame(players, center, playerNames, playerIndex);
 						if (getDoubleBonus(centerCard, playCard)) {
 							discards++;
 							draws++;
@@ -187,36 +223,51 @@ int main(void) {
 				}
 			}
 		}
+
+		// prompt them to play any additional cards
 		while (strcmp(scan, "continue") != 0) {
 			printf("Enter 'play' to play anothor card or 'continue' to continue your turn: ");
 			scanf("%s", scan);
+			
+			// obtain user input to make additionally plays
 			if (strcmp(scan, "play") == 0) {
 				printf("Choose a card on the center row: ");
 				scanf("%s %d", &centerCard.color, &centerCard.value);
 				printf("Enter a card to play: ");
 				scanf("%s %d", &playCard[0].color, &playCard[0].value);
 				playCard[1].value = 0;
-				//checks if the play is valid
-				if (checkPlay(center, players[playerIndex], centerCard, playCard)) {
-					//removes cards from player hand and center row,  
+				
+				// if the center card is a Wild # card, 
+				if (centerCard.value == 11) {
+					printf("Would you like to add another card? (Yes/No):");
+					scanf("%s", scan);
+				}
+
+				// check if the single match is valid 
+				if (checkPlay(center, players[playerIndex], centerCard, playCard) && (strcmp(scan, "Yes") != 0 && strcmp(scan, "yes") != 0)) {
+
+					// play the card from the player's hand and check for bonuses 
 					players[playerIndex] = deleteCard(players[playerIndex], playCard[0]);
 					center = deleteCard(center, centerCard);
 					centerNum--;
-					printGame(players, center, playerIndex);
+					printGame(players, center, playerNames, playerIndex);
 					if (strcmp(centerCard.color, playCard[0].color) == 0) {
 						discards++;
 					}
 				}
-				//if not, prompts for an additional card
+				/* if the single match is invalid or the user wishes to play two cards on a wildcard,
+				prompt for an additional card */
 				else if (playCard[0].value < centerCard.value) {
 					printf("Enter another card to play: ");
 					scanf("%s %d", &playCard[1].color, &playCard[1].value);
+
+					// if the double match is valid, play the cards and check for bonuses 
 					if (checkPlay(center, players[playerIndex], centerCard, playCard)) {
 						players[playerIndex] = deleteCard(players[playerIndex], playCard[0]);
 						players[playerIndex] = deleteCard(players[playerIndex], playCard[1]);
 						center = deleteCard(center, centerCard);
 						centerNum--;
-						printGame(players, center, playerIndex);
+						printGame(players, center, playerNames, playerIndex);
 						if (getDoubleBonus(centerCard, playCard)) {
 							discards++;
 							draws++;
@@ -229,32 +280,36 @@ int main(void) {
 				else {
 					printf("This is not a valid action\n");
 				}
-
-			}
+				if (players[playerIndex] == NULL || center == NULL) {
+					break;
+				}
+ 			}
 		}
 
-		//checks if there is a winner
+		// checks if there is a winner
 		if (players[playerIndex] == NULL) {
+			scores[playerIndex] = getPoints(players, playerIndex, playerNum);
 			break;
 		}
 
+		// draw up to two cards from the deck to the center row
 		for (i = centerNum; i < 2; i++) {
 			drawnCard = dealCard(deck, deckSize);
 			center = addCard(center, drawnCard);
 			centerNum++;
 		}
 
-		printGame(players, center, playerIndex);
+		printGame(players, center, playerNames, playerIndex);
 
-		//discard to the center row
+		// discard to the center row
 		i = 0;
-		while (i < discards) {
-			
+		while (i < discards) {	
 			printf("Enter a card to add to the center row: ");
 			scanf("%s %d", &centerCard.color, &centerCard.value);
 			temp = players[playerIndex];
 			while (temp != NULL) {
 
+				//if the inputed card exists, discard it to the center row
 				if (strcmp(temp->color, centerCard.color) == 0 && temp->value == centerCard.value) {
 					players[playerIndex] = deleteCard(players[playerIndex], centerCard);
 					center = addCard(center, centerCard);
@@ -272,6 +327,8 @@ int main(void) {
 			}
 
 		}
+
+		// deal out cards for the double match bonus
 		for (i = 0; i < draws; i++) {
 			for (j = 0; j < playerNum; j++) {
 				if (j != playerIndex) {
@@ -282,27 +339,24 @@ int main(void) {
 			}
 		}
 
-		//check if there is a winner 
+		// check if there is a winner 
 		if (players[playerIndex] == NULL) {
+			scores[playerIndex] = getPoints(players, playerIndex, playerNum);
 			break;
 		}
-		
-		//get the next player 
-		playerIndex = (playerIndex + 1) % playerNum;
-
-	}
-	
-	for (i = 0; i < playerNum; i++) {
-		if (i != playerIndex) {
-			printList(players[i]);
+		// get the next player 
+		else  {
+			playerIndex = (playerIndex + 1) % playerNum;
 		}
-
 	}
 
 	printf("\n+++++++++++++++++++++++++++++++++++++++++++\n");
-	printf("\n\n\n\t\tPlayer %d Wins\n", playerIndex + 1);
-	printf("\t\tScore: %d points\n\n\n", getPoints(players, playerIndex, playerNum));
-	printf("\n+++++++++++++++++++++++++++++++++++++++++++\n");
+	printf("\n\n\t%s Wins!\n", playerNames[playerIndex]);
+	printf("\tScores:" );
+	for (i = 0; i < playerNum; i++) {
+		printf("\n\t\t%s: \t%d", playerNames[playerIndex], scores[i]);
+	}
+	printf("\n\n\n+++++++++++++++++++++++++++++++++++++++++++\n");
 
 	return 0;
 }
